fix(fmp): address PR #560 review — named valuation type, immutable peer enrich, test import order, valuation error logging, extract provider type
